Update patch set 4
Patch Set 4: I would prefer that you didn't submit this

@Nathan, after playing with this code and seeing the impact it has on things like passing the download completed status to the main app (to update the progress bar for example), I would prefer this be abandoned.

I am not convinced that it solves a real end user problem and that the workarounds we will have to add for detecting when the main app is open or not etc. is worth the effort.  I am open to be convinced otherwise but don't see it at the moment.

Patch-set: 4