Update patch set 2
Patch Set 2:

Hrm DownloadProvider has some funky rules for the cache partition. What's the exception you're getting?

I'd prefer to not have a config like this. We should always use the cache partition if it's available. By available, I mean that we check the free space on it vs. the size of the download (it's part of the server response).

I suppose there are some masochistic users who like to have a stockpile of ROMs on their sdcard, but that's not a normal usage and if it's really needed we can add an option to the app to force /sdcard.

Patch-set: 2