Update patch set 2
Patch Set 2:

Looks like there's an easy trick..

http://stackoverflow.com/questions/4751791/downloadmanager-download-a-file-to-cache-partition

Just set destinationUri to null and it goes to cache.

I think we still have to deal with the "cache is full" issue anyway, because you can't copy the file if the cache is full. On Skyrocket /cache is tight- the ROM is 160MB and about 60MB is used by dexfiles in /cache, leaving us about 80MB with the file completed.

I think a reasonable way to do it is to:
1. Check if option to download to sdcard is enabled (off by default)
2. If off, check space on /cache and use it if there is enough free space.
3. Otherwise, download to sdcard.

Patch-set: 2