Update patch set 2
Patch Set 2:

For reference, similar change is here:
http://review.cyanogenmod.org/#/c/32910/

It would be nice if we could come up with a strategy that fixes both problems.  However, your problem is slightly different.

My problem in 32910 is that there are two different CWM's (cm10 and cm10.1 based) that map the sdcard's differently.

Patch-set: 2