Create change
Patch-set: 1
Change-id: I28facad38442c0e8d0b6fb0a70b71b340430e408
Subject: Adding a DashClock extension that shows the number of updates.  This can definitely be improved upon, but it is a good start.  This is nicer than a notification, and I have strived to make it look aesthetically pleasing on DashClock.
Branch: refs/heads/cm-10.1
Commit: 35bf21eb5034418fac3a7ba423bacbc99d391bc2