Create patch set 2
Uploaded patch set 2.

Patch-set: 2
Subject: Adding a DashClock extension that shows the number of updates.  This can definitely be improved upon, but it is a good start.  This is nicer than a notification, and I have strived to make it look aesthetically pleasing on DashClock.
Commit: cdc9f349458414d5a0ca3981c308686a8672a895