Create patch set 4
Uploaded patch set 4.

Patch-set: 4
Subject: DashClock extension that shows the number of updates. This can definitely be improved upon, but it is a good start.  This is nicer than a notification, and I have strived to make it look aesthetically pleasing on DashClock. Use Case: Friendly reminder for someone who likes to flash nighly builds as soon as they are uploaded. Possible future feature: add separate update check time to DahsClock extension, so it is not bound by the check time for the main app.
Commit: aa778155fd15a2d6e21bb82aee19021559916935