Update patch set 5
Patch Set 5:

UX people: This exists for the same reason there are builtin SMS and GMail extensions even though there are gmail and SMS notification, or a wildly popular battery extension even though there's a battery icon:

The lockscreen is used by people (myself included) as a summarized dashboard for "stuff" (which in my case includes calendar, pending texts, and battery. And yes, I'd like to see this one as well). The statusbar may have all of the same information, but it does not target the same usecase as the readable-from-a-distance lockscreen.

Patch-set: 5