Create patch set 7
Uploaded patch set 7.

Patch-set: 7
Subject: DashClock extension that shows the number of updates. This can definitely be improved upon, but it is a good start.  This is nicer than a notification, and I have strived to make it look aesthetically pleasing on DashClock. Use Case: Friendly reminder for someone who likes to flash nighly builds as soon as they are uploaded. Possible future feature: add separate update check time to DahsClock extension, so it is not bound by the check time for the main app.
Commit: 5a756561b7793375bced5212dbf18504fe691fbe