Update patch set 1
Patch Set 1:

With this patch there is no more notification, just a toast message when there are no new updates.

But there is one scenario that I think is weird, maybe you can explain;

When set to "All version (incl. Nightlies)" a notification message will show (with "On every boot" option) for OLDER nightlies.

See screenshots:
https://www.dropbox.com/s/njk02y6dl0u5ipy/Screenshot_2013-07-07-13-11-37.png
https://www.dropbox.com/s/njjs7c5gtj0mnu1/Screenshot_2013-07-07-13-10-55.png

This also happens when the nightly on the top has the "installed" status (I had that one installed before I built with this patch).

Should it not ignore older nightlies? I mean, not notify of those? Does not make much sense to me tbh.

Patch-set: 1