Update patch set 1
Patch Set 1:

Or maybe I am just confused and should just use the "New version" option instead.

Also, when having selected the "All versions (incl. Nightlies)" option and then changing the "Check for updates" option - for example, from 'on every boot' to 'manual' - a notification is shown saying "new updates are available", but that notification is not sticking, it clears itself immediately.

I don't know if this is by design or not....I think a toast message would be sufficient in this case.

Patch-set: 1