Update patch set 3
Patch Set 3:

Danny, IMHO if the updater app is open and the download completes, the install prompt should show immediately - the user is after all in the process of updating and waiting for the download to complete - showing a notification and adding another action for the user to perform is redundant.   If the app is closed and the user is in another app, just showing the  notification makes sense provided it has the reboot and install action associated as well.

Patch-set: 3