Update patch set 3
Patch Set 3: -Code-Review

Well, not showing the notification if the activity is already active makes sense to me; I just don't see how to check the foreground activity from a broadcast receiver.

I've left out the 'Install' button from the notification so far, as I thought we'd want to show the dialog with the '... requires a compatible recovery...' hint first. If that's not required, it's easy to add that button (just need to shuffle a bit of code around).

Patch-set: 3