Update patch set 1
Patch Set 1:

Yes, I had already downloaded the file once completely before but deleted it, since I wanted to test the text that is shown when downloading is finished (for new translation).
After that I started a new download.

And yes, that new download had started and it failed while in a call.

So both your assumptions are right

Patch-set: 1