Update patch set 4
Patch Set 4:

Yeah, I was being lazy.  I'm going to remove Ion today.

DownloadReceiver isn't doing any heavy lifting.  It is just creating a request to send to download manager, and checking for a possible incremental (asynchronously, using Ion).  If I replace Ion, I'll do so using an async task, which imo should still be fine to run on the main thread.

If you *really* want to move this into a service, let me know, that'll take a little bit more time and is probably outside of the scope of this patch.

Patch-set: 4