Update patch set 1
Patch Set 1:

Also, when a download IS complete, pressing the Restart and Update button seems to do nothing. A user trying it for the first time does not know that it needs to wait till the device restarts itself.
On my i9i00 it takes about 10 seconds. In those 10 seconds, the device is frozen. 

Maybe some change can be made to show the user that the device is rebooting? A normal reboot shows a spinning icon, we need that here too.

Patch-set: 1