Update patch set 6
Patch Set 6: Code-Review+1

I made the UpdateInfo that's passed as an extra in the intent of DownloadService a member variable so that I can fetch information from it later in the now implemented onResponse -- since we're only allowing a single download at a time it shouldn't be a problem. Either way, IntentService handles each intent passed sequentially so it shouldn't really matter.

Patch-set: 6