Update patch set 8
Patch Set 8: Verified+1 Code-Review+1

Talked to Soyars about it, should be available in the future. In the mean time, I think this should be merged as-is with the changelog fetch being done via an explicit asynctask in http://review.cyanogenmod.org/#/c/70792/.

Whenever we do have the ability to fetch compressed data for changelogs from the server, I'll push up a patch to add a header for it.

Patch-set: 8
Reviewer: Adnan Begovic <9813@1ec6b3db-b2c1-4fa4-84a2-4c7efe89ba71>
Label: Verified=+1