Update patch set 3
Patch Set 3: Code-Review+2

Originally I thought it might be beneficial for the end user to have downgrade available to them if on the same API level (which this didn't take into account), however, that option requires putting a lot of unnecessary cruft like dialogs, warnings, and uglyness in the UI and placing enough trust that the end user understands that a downgrade without a wipe would be disastrous.


So, +2 for me. If ever a user would want a downgrade, they'll have to do it manually. This should also hopefully resolve the most automated bugreports related to database downgrade issues whenever we merge in CAF nightmares (Psychoi3oy will be ecstatic).

Patch-set: 3
Label: Code-Review=+2