Update patch set 3
Patch Set 3: Verified-1

Personally, codestyle cleanup can cause a lot of harm. When if ever you're doing a git blame on a specific piece of code to see when it was introduced, a vast commit like this will skew the perceived history of any code it touches. I'd prefer just the removal of dashclock support and autobackup, vs the mass cleanup.

Patch-set: 3