Update patch set 3
Patch Set 3:

FYI - this commit has caused a problem for cmupdater driven updates ever since it was merged.

I originally thought the problem was TWRP but I've had it happen with a mid-jan cm-11.0 recovery too.

Revert this and all is well again.  Workarounds are sideload or download with cmupdater but then reboot manually into recovery to install.

Trigger: download update via cmupdater (or adb push directly into cmupdater dir as I Do), reboot to upgrade and it bombs out when trying to install the zip (unknown volume for path [@/cache/recovery/block.map]).

Device type doesn't seem to matter.

It's not clear to me why this commit is causing the problem and it may be recoveries that should ideally be fixed but the fact remains - CM is currently getting bad PR because of this.

https://www.google.com/search?q=unknown+volume+for+path+%22%5B%40%2Fcache%2Frecovery%2Fblock.map%5D%22

Patch-set: 3
Reviewer: Sam Mortimer <3075@1ec6b3db-b2c1-4fa4-84a2-4c7efe89ba71>
Label: Verified=0