Update patch set 2
Patch Set 2:

OK. I am not very proud of this, but here we go...

https://github.com/CyanogenMod/android_device_bn_acclaim/blob/cm-12.0/overlay/aosp/frameworks/base/core/res/res/xml/storage_list.xml

The partition layout of the acclaim is quite weird. A "huge" ext4 12GB partition and a small vfat 1GB one. So I decided at some point it might be better to force people to put a sdcard in and use this as a primary storage instead of 1GB partition.

But OK. I guess I will have to do something about this in the end (repartition or flip the partitions around - i.e. use 1GB for /data and 14GB for /data/media, or just use 12 GB for emulated storage and use 1GB as 2nd storage and external SD card as "USB storage/2nd SD card" ).

Patch-set: 2