Update patch set 1
Patch Set 1:

Danny, the notification shows on the Wearable device with or without this patch. This simply overrides the wearable install action, and swaps in the icon that provides a better UX (white icon on blue action target). The same can/should be done for the download notification, since it had the same material teal on blue last night.

Patch-set: 1