Update patch set 2
Patch Set 2:

This is more important for CyanogenMod since the builds are signed with the test keys. On CopperheadOS, it only prevents evil apps with storage access from blocking updates by corrupting them as the signature gets verified by the recovery image. The more significant change was moving to uncrypt to fix updates with encryption + AOSP recovery (copying to /cache was a workaround in the past, but it's too small on newer devices) but CyanogenMod ended up doing that anyway. I wasn't sure if you would want to make these changes so I didn't submit them myself.

Patch-set: 2