Update patch set 2
Patch Set 2:

Beyond the topic of this specific patch:

I think the best long-term approach for CyanogenMod would be signing the builds (snapshots AND nightlies) and adding the test releasekey as an extra recovery signature to allow users to sideload. The /system/priv-app directory can be used to gain most platform permissions rather than the platform key. AFAIK, anything specific to the platform key could just be extended to privileged system applications. The compatibility check could be altered to allow updates from the current test keys builds (if the simple recovery currently enforces the test key, that could be removed in an update as an incremental step and then added back with the real key).

Even if users are using a recovery unable to verify the signature (TWRP), it can at least be verified by the updater (there's a RecoverySystem verifyPackage API) and the private directory will make it relatively safe. HTTPS without a pinned key is much weaker. It also defends against a server hack if the builds are signed on a separate machine and then uploaded there.

Patch-set: 2
Label: Code-Review=0