Update patch set 2
Patch Set 2:

@Joey
So true. Especially for nighties, the most important feature of the CMUpdater is exactly that it is possible to reflash the previous release, if something goes wrong. For example when there is a bug in the kernel.

Patch-set: 2
Reviewer: Henning Nielsen Lund <3848@1ec6b3db-b2c1-4fa4-84a2-4c7efe89ba71>
Label: Verified=0