Update patch set 2
Patch Set 2:

You mean /data/data, right? It doesn't use /cache.

I don't have much interest in pushing for this. Signing the builds and performing updates securely would be good, but the security issues in CyanogenMod run much deeper than that. CyanogenMod wouldn't have the power to secure most of the supported devices even if security was a top priority since the vendors don't provide security fixes for the proprietary components. It would be nice if it was on par with the security of AOSP on Nexus devices, but I don't think that's an obtainable goal - even getting close is unrealistic. There would be many sacrifices required, and there's no willpower to even make *this* change...

Patch-set: 2