Update patch set 3
Patch Set 3:

You'll either need to set the target API level to 23 as we do or come up with an alternate fix for this. The CyanogenMod fix doesn't work with the internal storage directory, so I reverted it. Might just need to add a field to keep track of the file. It's annoying that they removed this for security and it gets in the way of a real security fix...

Patch-set: 3