Update patch set 4
Patch Set 4:

Is there any reason not to check both the signature and the md5 other than not wanting to provide it with the server? If the signature intact is that just as good to make sure the integrity of the file hasn't been compromised?

Patch-set: 4
Reviewer: Matthew Garbett <9738@1ec6b3db-b2c1-4fa4-84a2-4c7efe89ba71>
Label: Verified=0