Update patch set 10
Patch Set 10:

Upon further investigation, "verifyPackage" is working correctly in the Updater application.  The actual problem is that when Updater calls the DownloadManager, the ota.zip becomes corrupted. The download appears to be successful, but it is not.

01-21 02:34:05.200 13173 18197 V DownloadService: Downloading full zip
01-21 02:34:05.295 13227 18200 D DownloadManager: [292] Starting
01-21 02:40:43.908 13227 18200 D DownloadManager: [292] Finished with status SUCCESS
01-21 02:41:11.646 13227 13239 V DownloadManager: Deleting /data/data/com.android.providers.downloads/cache/lineage-14.1-20170121-NIGHTLY-v500.zip via provider delete

If I download the same update via the http interface with a separate web browser, then the ota.zip is not corrupted.

I also noticed that the Updater checks for new updates while the updater is currently busy downloading a new ota.  It also checks for a new update right after the download is finished.  I have it set to manual.

Patch-set: 10