Update patch set 10
Patch Set 10:

With the commit above, the download appeared to be finished in the Updater, then logcat said it was downloaded successful (but it was not finished).  Then the download started up again and finished the last 10% of the file and verifyPackage then said it was successful.

I believe during my test, I also had an additional wait statement right before the verifyPackage call.  With the wait statements, it appears that the download will resume and finish before it is checked so then it works as expected.

Patch-set: 10