Update patch set 1
Patch Set 1:

One thing that's still not quite right is it doesn't seem to always apply the style to the UpdatePreference. I seem to be able to reproduce that fairly often when killing the activity and starting it new, then trying to hit download. The download manager starts, the runnable is runnable-ing, its getting progress, but the style of the preference isn't set to STYLE_DOWNLOADING. Hell, I even tried setting it within the runnable and it didn't make a difference.

*scratches head*

Patch-set: 1