Update patch set 2
Patch Set 2:

It's entirely possible that in this case FileChannel.size() returns the actual size of the file, but I see no guarantees of that in the documentation. In any case, we already have a ParcelFileDescriptor object and getStatSize() guarantees that what we get is the size of the file, so I'd switch to that, even if in the end this won't make any difference.

Patch-set: 2
Label: Verified=0