Update patch set 1
Patch Set 1:

The main problem is that we can get a null pointer until a build has been downloaded. After the first successful, we either get a proper string or an empty string (the pref is not removed, but overridden, see few lines below), which is equally bad.

I completely changed the file name handling with another commit by the way, see 166737.

Patch-set: 1
Label: Code-Review=0