Update patch set 2
Patch Set 2:

> The main problem is that we can get a null pointer until a build
 > has been downloaded. After the first successful, we either get a
 > proper string or an empty string (the pref is not removed, but
 > overridden, see few lines below), which is equally bad.
 > 
 > I completely changed the file name handling with another commit by
 > the way, see 166737.

I'm assuming you're referring to:
mPrefs.edit().putString(Constants.DOWNLOAD_NAME, "").commit();
in which case, I'm confused, because we already have a copy of what Constants.DOWNLOAD_MODE was before we overrode it. Or are you referring to that if we override it and go through this method again, we get an empty string? In that case, I can add a check to break out if it's empty, seems like less changes than redoing the handling

Patch-set: 2
Reviewer: Paul Keith <11801@1ec6b3db-b2c1-4fa4-84a2-4c7efe89ba71>
Label: Verified=0