Don't allow new downloads while verifying another download
After the file has been downloaded, it's copied from the
DownloadManager cache into our private storage and then verified.
Both these operations require some time to complete. If the main
activity is not up when these operations are started, we incorrectly
show the "Install" button even if the file is yet to be verified. Keep
track of the completing downloads and add some visual feedback to let
the user know about the ongoing operations. To not overcomplicate the
code, don't allow new downloads while another one is being completed.

Change-Id: I6c0dbbf50474a9cfad919ebc372c00615ce8d6de