Update patch set 6
Patch Set 6:

> I'm not talking about the code itself, but rather that you remove
 > the error message, which is nice to have when stuff goes wrong, to
 > nicely explain what's up

I got that. What I was trying to say is that we should never hit the error now.

It's pretty much guaranteed that both the hasExtras() conditions are true. The error there is more of a hint for developers.

If we really want to display something, that should be done from handleDownloadComplete(), but I think it's not required. These two prefs are now accessed from only two different places now: in one we set them (when we start the download) and in the other we read and clear them (when the download is complete and we yet have to verify the zip). Before this change it was possible to have one of the prefs not properly set because we accessed them in different moments from different places (e.g. it was possible to have the name overridden before onHandleIntent()).

Patch-set: 6
Label: Verified=0