Update patch set 3
Patch Set 3:

Because of the missing 'status != DownloadManager.STATUS_SUCCESSFUL', the "Install" button appeared as soon as the zip was removed from DownloadManager, even if it wasn't verified.

I also dropped the sharedpreference and switched to a name suffix to track the completing downloads. I think it's simpler like this.

Patch-set: 3
Label: Verified=0