Fix dates of downloaded and installed builds
With commit 30b0384ba125 ("Move properties parsing logic inside Utils")
it was assumed that getDate() returns a proper timestamp for all the
files, but this is only true for builds not yet downloaded. For all
the other builds the returned value is 0, so, if the timestamp is 0,
try to get the date from the filename.

Change-Id: Id61fc9a45f206f5ce8020ac845610c2c0fb9a90d